Important: Major Update
This patch fixes several major & minor bugs and includes lots of
features for developers and users.

*Module Installer Patched - The Installer module was found to be buggy
and was fixed & secured and now is ready to be used.

*dbSQL.sql - The SQL files which were provided were incorrect and needed
to be updated so that is done, if any case the installer plays up, run
the this file in PMA.

*errorMessage(); - Show a error layout and give custom messages.

*successMessage(); - Show a success layout and give custom messages.

*getMoney(); - Allows developers to get users money from the users
class, and shows the money that the user has.

*addMoney(); - Easily allows developers to add money to player that they
want.

*minusMoney(); - Easily allows developers to minus money to player that
they want.

*Money - Added some sort of currency to the game.

*Module guessthenumber - Allows you to test the new implemented money
field.